fix login/logout
