+++ src/hooks/useUser/index.js
@@ -7,7 +7,6 @@ const useUser = () => {
   const hash = window.location.hash;
 
   useEffect(() => {
-    console.log(token);
     const fetchData = async () => {
       const user = await getUser(token);
       setUserId(user?.id);
@@ -33,7 +32,7 @@ const useUser = () => {
   const logout = () => {
     setToken(null);
     setUserId(null);
-    localStorage.removeItem("token");
+    window.localStorage.removeItem("token");
     window.location.reload();
   };
